PowerTS_repo/requirements.txt

numpy
pandas
pyyaml
torch>=2.0.0
